Split native app shell DTO module
